Remove SequenceLength and use tabs for sequence length un/reviewed selection.
